requirements.txt
asgiref==3.7.2
Django==5.0.1
django-extensions==3.2.3
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.4
